backend/requirements.txt
fastapi
uvicorn
pydantic
requests
newspaper3k
sumy
sqlalchemy
aiosqlite
